Capturing working, need ko
Capturing is working!! Tested with 9x9 board and all seems to be working fine apart from KO.

Not sure how to implement ko...because it is a very specific position.
Maybe by checking if last move (which would capturing game state and storing it in a history of the game)
Need to implement PASS button.